add reverse-engineered source of srampatch
this one was surprisingly simple to reverse engineer.
result compiles, and a bug causing double-free which results in a
crash on linux and an empty output file on window "fixed" - in that
no crash is happening anymore. the fix probably needs some more work
in order to propagate the modified filesize to top-level.
a follow-up commit will make my fix optional, for total transparency.

apart from mentioned bug, the reversed code seems to work as well
as the original .exe in preliminary tests.

the bug mentioned only affects a handful roms using the first EEPROM
chip. supercard team was apparently aware of the bug and therefore
do a funny dance in the Supercardsd.exe:

when user clicks "Add" button, `srampatch.exe 128 temp/xxx.gba temp/yyy.gba`
is executed via ShellExecuteA, where xxx and yyy are random hexadecimal names.
if yyy.gba is of filesize 0, the game is marked as affected by the bug
and will then be patched via gep.exe instead when the user selects the
"Enable saver patch" option.