srampatch.c: properly fix eeprom type 11x bug
with an input rom that has all 0xff-padding truncated, this now
produces the same result as gep.exe, with the only difference that
gep always appends a full 256 bytes and pads the rest with 0x00's.

in order to fix the issue, 2 new global variables were introduces
that are written to by the eeprom func, and checked before writing
the rom and freeing the data.