srampatch.c: produce identical result to gep.exe for eeprom v11x
well, not with vanilla gep.exe, but with how Supercardsd.exe uses
it:

first, all trailing 0xff (aligned to 4) are truncated.
then, 512 zero bytes are appended. this is used by gep.exe so it
has space for the eeprom code.

srampatch.c now does the same thing, except that the input needs to
be truncated already, and no zero-padding is necessary, as it does
it itself.